requirements.txt
django
djangorestframework
django-cors-headers
yfinance
pandas
